+++ src/@types/index.d.ts
@@ -113,6 +113,7 @@ declare global {
              * 成立した取引結果
              */
             transactionResult?: cinerinoapi.factory.transaction.placeOrder.IResult & {
+                code?: string;
                 paymentNo: string;
                 printToken: string;
             };