+++ src/app/controllers/customer/reserve.ts
@@ -6,7 +6,7 @@ import * as conf from 'config';
 import * as createDebug from 'debug';
 import { NextFunction, Request, Response } from 'express';
 import { BAD_REQUEST, CONFLICT, INTERNAL_SERVER_ERROR, NOT_FOUND, TOO_MANY_REQUESTS } from 'http-status';
-import * as jwt from 'jsonwebtoken';
+// import * as jwt from 'jsonwebtoken';
 import * as moment from 'moment-timezone';
 
 import reservePaymentCreditForm from '../../forms/reserve/reservePaymentCreditForm';
@@ -442,15 +442,15 @@ export async function confirm(req: Request, res: Response, next: NextFunction): 
                 }
 
                 // 印刷トークン生成
-                const reservationIds =
-                    transactionResult.order.acceptedOffers.map((o) => (<cinerinoapi.factory.order.IReservation>o.itemOffered).id);
-                const printToken = await createPrintToken(reservationIds);
+                // const reservationIds =
+                //     transactionResult.order.acceptedOffers.map((o) => (<cinerinoapi.factory.order.IReservation>o.itemOffered).id);
+                // const printToken = await createPrintToken(reservationIds);
 
                 // 購入結果セッション作成
                 (<Express.Session>req.session).transactionResult = {
                     ...transactionResult,
                     code,
-                    printToken,
+                    // printToken,
                     paymentNo: transactionResult.order.confirmationNumber
                 };
 
@@ -518,26 +518,26 @@ export type IPrintToken = string;
 /**
  * 印刷トークン対象(予約IDリスト)インターフェース
  */
-export type IPrintObject = string[];
+// export type IPrintObject = string[];
 
 /**
  * 予約印刷トークンを発行する
  */
-async function createPrintToken(object: IPrintObject): Promise<IPrintToken> {
-    return new Promise<IPrintToken>((resolve, reject) => {
-        const payload = {
-            object: object
-        };
+// async function createPrintToken(object: IPrintObject): Promise<IPrintToken> {
+//     return new Promise<IPrintToken>((resolve, reject) => {
+//         const payload = {
+//             object: object
+//         };
 
-        jwt.sign(payload, <string>process.env.TTTS_TOKEN_SECRET, (jwtErr, token) => {
-            if (jwtErr instanceof Error) {
-                reject(jwtErr);
-            } else {
-                resolve(token);
-            }
-        });
-    });
-}
+//         jwt.sign(payload, <string>process.env.TTTS_TOKEN_SECRET, (jwtErr, token) => {
+//             if (jwtErr instanceof Error) {
+//                 reject(jwtErr);
+//             } else {
+//                 resolve(token);
+//             }
+//         });
+//     });
+// }
 
 /**
  * 予約完了
@@ -570,8 +570,8 @@ export async function complete(req: Request, res: Response, next: NextFunction):
         res.render('customer/reserve/complete', {
             order: transactionResult.order,
             reservations: reservations,
-            paymentNo: transactionResult.paymentNo,
-            printToken: transactionResult.printToken
+            paymentNo: transactionResult.paymentNo
+            // printToken: transactionResult.printToken
         });
     } catch (error) {
         next(new Error(req.__('UnexpectedError')));