+++ dst/app/controllers/customer/reserve.js
@@ -9,7 +9,7 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
     });
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.complete = exports.createEmail = exports.confirm = exports.profile = exports.tickets = exports.performances = exports.changeCategory = exports.start = void 0;
+exports.complete = exports.createEmail = exports.confirm = exports.profile = exports.tickets = exports.performances = exports.changeCategory = exports.start = exports.CODE_EXPIRES_IN_SECONDS = void 0;
 /**
  * 予約コントローラー
  */
@@ -17,13 +17,13 @@ const cinerinoapi = require("@cinerino/sdk");
 const conf = require("config");
 const createDebug = require("debug");
 const http_status_1 = require("http-status");
-const jwt = require("jsonwebtoken");
 const moment = require("moment-timezone");
 const reservePaymentCreditForm_1 = require("../../forms/reserve/reservePaymentCreditForm");
 const reservePerformanceForm_1 = require("../../forms/reserve/reservePerformanceForm");
 const session_1 = require("../../models/reserve/session");
 const reserveBaseController = require("../reserveBase");
 const reserve_1 = require("../../factory/reserve");
+exports.CODE_EXPIRES_IN_SECONDS = 8035200; // 93日
 const debug = createDebug('ttts-frontend:controller:customerReserve');
 const reserveMaxDateInfo = conf.get('reserve_max_date');
 const authClient = new cinerinoapi.auth.ClientCredentials({
@@ -32,6 +32,10 @@ const authClient = new cinerinoapi.auth.ClientCredentials({
     clientSecret: process.env.API_CLIENT_SECRET,
     scopes: [],
     state: ''
+});
+const orderService = new cinerinoapi.service.Order({
+    endpoint: process.env.CINERINO_API_ENDPOINT,
+    auth: authClient
 });
 const placeOrderTransactionService = new cinerinoapi.service.transaction.PlaceOrder4ttts({
     endpoint: process.env.CINERINO_API_ENDPOINT,
@@ -58,7 +62,6 @@ function start(req, res, next) {
         try {
             // 購入結果セッション初期化
             delete req.session.transactionResult;
-            delete req.session.printToken;
             const reservationModel = yield reserveBaseController.processStart(req);
             reservationModel.save(req);
             // パフォーマンス選択へ遷移
@@ -229,6 +232,7 @@ exports.tickets = tickets;
  */
 // tslint:disable-next-line:max-func-body-length
 function profile(req, res, next) {
+    var _a;
     return __awaiter(this, void 0, void 0, function* () {
         try {
             const reservationModel = session_1.default.FIND(req);
@@ -317,7 +321,7 @@ function profile(req, res, next) {
             if (paymentAccepted !== undefined) {
                 const creditCardPaymentAccepted = paymentAccepted.find((p) => p.paymentMethodType === cinerinoapi.factory.paymentMethodType.CreditCard);
                 if (creditCardPaymentAccepted !== undefined) {
-                    gmoShopId = creditCardPaymentAccepted.gmoInfo.shopId;
+                    gmoShopId = (_a = creditCardPaymentAccepted.gmoInfo) === null || _a === void 0 ? void 0 : _a.shopId;
                 }
             }
             res.render('customer/reserve/profile', {
@@ -382,12 +386,34 @@ function confirm(req, res, next) {
                             }
                         }
                     });
+                    const order = transactionResult.order;
                     debug('transacion confirmed. orderNumber:', transactionResult.order.orderNumber);
-                    // 印刷トークン生成
-                    const reservationIds = transactionResult.order.acceptedOffers.map((o) => o.itemOffered.id);
-                    const printToken = yield createPrintToken(reservationIds);
+                    // 注文承認
+                    let code;
+                    try {
+                        // まず注文作成(非同期処理が間に合わない可能性ありなので)
+                        yield orderService.placeOrder(Object.assign({ orderNumber: order.orderNumber }, {
+                            confirmationNumber: order.confirmationNumber
+                        }));
+                        debug('order placed', order.orderNumber);
+                        const authorizeOrderResult = yield orderService.authorize({
+                            object: {
+                                orderNumber: order.orderNumber,
+                                customer: { telephone: order.customer.telephone }
+                            },
+                            result: {
+                                expiresInSeconds: exports.CODE_EXPIRES_IN_SECONDS
+                            }
+                        });
+                        code = authorizeOrderResult.code;
+                        debug('order code published', code);
+                    }
+                    catch (error) {
+                        // tslint:disable-next-line:no-console
+                        console.error(error);
+                    }
                     // 購入結果セッション作成
-                    req.session.transactionResult = Object.assign(Object.assign({}, transactionResult), { printToken, paymentNo: transactionResult.order.confirmationNumber });
+                    req.session.transactionResult = Object.assign(Object.assign(Object.assign({}, transactionResult), { paymentNo: transactionResult.order.confirmationNumber }), (typeof code === 'string') ? { code } : undefined);
                     // 購入フローセッションは削除
                     session_1.default.REMOVE(req);
                     res.redirect('/customer/reserve/complete');
@@ -435,26 +461,6 @@ function createEmail(reservationModel, res) {
 }
 exports.createEmail = createEmail;
 /**
- * 予約印刷トークンを発行する
- */
-function createPrintToken(object) {
-    return __awaiter(this, void 0, void 0, function* () {
-        return new Promise((resolve, reject) => {
-            const payload = {
-                object: object
-            };
-            jwt.sign(payload, process.env.TTTS_TOKEN_SECRET, (jwtErr, token) => {
-                if (jwtErr instanceof Error) {
-                    reject(jwtErr);
-                }
-                else {
-                    resolve(token);
-                }
-            });
-        });
-    });
-}
-/**
  * 予約完了
  */
 function complete(req, res, next) {
@@ -474,12 +480,7 @@ function complete(req, res, next) {
             });
             // チケットを券種コードでソート
             sortReservationstByTicketType(reservations);
-            res.render('customer/reserve/complete', {
-                order: transactionResult.order,
-                reservations: reservations,
-                paymentNo: transactionResult.paymentNo,
-                printToken: transactionResult.printToken
-            });
+            res.render('customer/reserve/complete', Object.assign({ order: transactionResult.order, reservations: reservations, paymentNo: transactionResult.paymentNo }, (typeof transactionResult.code === 'string') ? { code: transactionResult.code } : undefined));
         }
         catch (error) {
             next(new Error(req.__('UnexpectedError')));
@@ -526,12 +527,11 @@ function processFixGMO(reservationModel, req) {
                 const gmoTokenObject = JSON.parse(req.body.gmoTokenObject);
                 const amount = reservationModel.getTotalCharge();
                 // クレジットカードオーソリ取得
-                debug('creating credit card authorizeAction...');
                 const action = yield paymentService.authorizeCreditCard({
                     object: {
-                        typeOf: cinerinoapi.factory.paymentMethodType.CreditCard,
+                        typeOf: cinerinoapi.factory.action.authorize.paymentMethod.any.ResultType.Payment,
+                        paymentMethod: cinerinoapi.factory.chevre.paymentMethodType.CreditCard,
                         amount: amount,
-                        // tslint:disable-next-line:no-suspicious-comment
                         method: '1',
                         creditCard: gmoTokenObject
                     },