+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - update @cinerino/sdk
 - チケット印刷URLをprintByOrderNumberに変更
+- 購入完了ページ表示前に注文承認処理を追加
 
 ### Deprecated